refactor: update imports after changing dtos packages (#11)
